Remove extension, rename `ViewRenderer.php` to `TemplateRenderer.php`, update docs
